fix(tests): fix Float32->Float64 type mismatch in assert_close_float
Convert Float32 literal arguments to Float64 when calling assert_close_float,
which expects Float64 parameters. Tests on lines 90 and 97 were passing Float32
values, causing compilation to fail with --Werror.

Closes #4520

+++ tests/shared/testing/test_assertions_float.mojo
@@ -87,14 +87,18 @@ fn test_assert_equal_float_fails() raises:
 
 fn test_assert_close_float_passes() raises:
     """Test assert_close_float with values within tolerance. Closes #4096."""
-    assert_close_float(Float32(1.0), Float32(1.0001), rtol=1e-2, atol=1e-3)
+    assert_close_float(
+        Float64(1.0), Float64(1.0001), rtol=1e-2, atol=1e-3
+    )
 
 
 fn test_assert_close_float_fails() raises:
     """Test assert_close_float raises for distant values. Closes #4096."""
     var failed = False
     try:
-        assert_close_float(Float32(1.0), Float32(2.0), rtol=1e-5, atol=1e-5)
+        assert_close_float(
+            Float64(1.0), Float64(2.0), rtol=1e-5, atol=1e-5
+        )
     except:
         failed = True
     assert_true(
